Make routes modular
