+++ service/src/routes/index.ts
@@ -1,0 +1,5 @@
+import user from './user';
+
+export default {
+  user
+};
